requirements.txt
streamlit==1.32.0
openai-whisper==20231117
ffmpeg-python
